+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Easier to display items in a list/grid view from your controllers directly or handling state internally with support for pagination.